packages/backend/requirements.txt
fastapi
uvicorn
pydantic[email]
python-dotenv
supabase
psycopg2-binary
jinja2
numpy
scikit-learn
pandas
requests
aioredis
pylint
autopep8
isort
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
celery
redis
prometheus-client
prometheus-fastapi-instrumentator
slowapi
sendgrid
exponent-server-sdk
fpdf2
